refactor(configurator): split field update domains
